Adds state to storefront orders.
State is a new field on the Order of the enum type OrderState.
By saving regularly between microservice calls the storefront can control the flow.
This is orchestrated by the CartService, but we could have another class between the controller and the CartService doing that (retries, cancel etc.).

CartDto received a new field orderId (ignored on the responses to the client via @JsonIgnore annotation).
This allows for the failback method of CartService to retrieve the saved Order instance.

Some toString methods were created to improve logging.