Adds security and OAuth to the three main services
Protects POST requests on all applications.
Configures @EnableResourceServer and configures the corresponding user-inf-uri.

Adds a final version of the Insomnia Requests with POST requests already configured to use the authentication app.

Last but not least (in fact, that's crucial), configures a bean in the flowerstore application to get the authentication headers and add them to a RequestTemplate that is used by the FeignClients when integrating with other microservices.